Grass
Added Frigid Grass
Various Grasses now have Y offset too

+++ src/main/java/net/hibiscus/naturespirit/client/MapleLeavesParticle.java
@@ -12,8 +12,7 @@ import net.minecraft.client.particle.SpriteBillboardParticle;
 import net.minecraft.client.particle.SpriteProvider;
 import net.minecraft.client.world.ClientWorld;
 
-@Environment(EnvType.CLIENT)
-public class MapleLeavesParticle extends SpriteBillboardParticle {
+@Environment(EnvType.CLIENT) public class MapleLeavesParticle extends SpriteBillboardParticle {
    private static final float field_43372 = 0.0025F;
    private static final int field_43373 = 300;
    private static final int field_43366 = 300;
@@ -26,9 +25,9 @@ public class MapleLeavesParticle extends SpriteBillboardParticle {
    protected MapleLeavesParticle(ClientWorld world, double x, double y, double z, SpriteProvider spriteProvider) {
       super(world, x, y, z);
       this.setSprite(spriteProvider.getSprite(this.random.nextInt(12), 12));
-      this.field_43369 = (float)Math.toRadians(this.random.nextBoolean() ? -30.0D : 30.0D);
+      this.field_43369 = (float) Math.toRadians(this.random.nextBoolean() ? -30.0D : 30.0D);
       this.field_43370 = this.random.nextFloat();
-      this.field_43371 = (float)Math.toRadians(this.random.nextBoolean() ? -5.0D : 5.0D);
+      this.field_43371 = (float) Math.toRadians(this.random.nextBoolean() ? -5.0D : 5.0D);
       this.maxAge = 300;
       this.gravityStrength = 7.5E-4F;
       float f = this.random.nextBoolean() ? 0.133333F : 0.19999999995F;
@@ -45,30 +44,30 @@ public class MapleLeavesParticle extends SpriteBillboardParticle {
       this.prevPosX = this.x;
       this.prevPosY = this.y;
       this.prevPosZ = this.z;
-      if (this.maxAge-- <= 0) {
+      if(this.maxAge-- <= 0) {
          this.markDead();
       }
 
-      if (!this.dead) {
-         float f = (float)(300 - this.maxAge);
+      if(!this.dead) {
+         float f = (float) (300 - this.maxAge);
          float g = Math.min(f / 300.0F, 1.0F);
-         double d = Math.cos(Math.toRadians((double)(this.field_43370 * 60.0F))) * 1.85D * Math.pow((double)g, 1.25D);
-         double e = Math.sin(Math.toRadians((double)(this.field_43370 * 60.0F))) * 1.85D * Math.pow((double)g, 1.25D);
+         double d = Math.cos(Math.toRadians(this.field_43370 * 60.0F)) * 1.85D * Math.pow(g, 1.25D);
+         double e = Math.sin(Math.toRadians(this.field_43370 * 60.0F)) * 1.85D * Math.pow(g, 1.25D);
          this.velocityX += d * 0.0024999999441206455D;
          this.velocityZ += e * 0.0024999999441206455D;
-         this.velocityY -= (double)this.gravityStrength;
+         this.velocityY -= this.gravityStrength;
          this.field_43369 += this.field_43371 / 20.0F;
          this.prevAngle = this.angle;
          this.angle += this.field_43369 / 20.0F;
          this.move(this.velocityX, this.velocityY, this.velocityZ);
-         if (this.onGround || this.maxAge < 299 && (this.velocityX == 0.0D || this.velocityZ == 0.0D)) {
+         if(this.onGround || this.maxAge < 299 && (this.velocityX == 0.0D || this.velocityZ == 0.0D)) {
             this.markDead();
          }
 
-         if (!this.dead) {
-            this.velocityX *= (double)this.velocityMultiplier;
-            this.velocityY *= (double)this.velocityMultiplier;
-            this.velocityZ *= (double)this.velocityMultiplier;
+         if(!this.dead) {
+            this.velocityX *= this.velocityMultiplier;
+            this.velocityY *= this.velocityMultiplier;
+            this.velocityZ *= this.velocityMultiplier;
          }
       }
    }
